Template to translate world tags
All those tags are empty, but the number of enemies, friends, complication, things and places should be exact. This should speed up the translation effort : simple copy and paste from the translation google doc, and remove the extra line once you're done. 

The extra line between every tag is there to help spot mistakes quickly : if there is a mistake in one tag, the translation won't line up, and you'll know the mistake is in the tag you just filled. Less wasted effort this way.